Split navigation soak runner and command wiring into dedicated modules
