+++ src/yoyopod/cli/pi/navigation/__init__.py
@@ -1,0 +1,14 @@
+"""Target-hardware navigation and idle soak helpers."""
+
+from __future__ import annotations
+
+from yoyopod.cli.pi.navigation.command import run_navigation_soak
+from yoyopod.cli.pi.navigation.stats import NavigationSoakFailure, NavigationSoakStats
+from yoyopod.cli.pi.navigation.runner import NavigationSoakRunner
+
+__all__ = [
+    "NavigationSoakFailure",
+    "NavigationSoakStats",
+    "NavigationSoakRunner",
+    "run_navigation_soak",
+]
