+++ src/yoyopod/cli/pi/validate.py
@@ -13,7 +13,7 @@ import typer
 from yoyopod.audio.test_music import DEFAULT_TEST_MUSIC_TARGET_DIR
 from yoyopod.cli.common import REPO_ROOT, configure_logging, resolve_config_dir
 from yoyopod.cli.pi.lvgl import soak as run_lvgl_soak
-from yoyopod.cli.pi.navigation import run_navigation_soak
+from yoyopod.cli.pi.navigation.command import register_navigation_command
 from yoyopod.cli.pi.smoke import (
     CheckResult,
     _display_check,
@@ -390,80 +390,4 @@ def stability(
     )
 
 
-@validate_app.command()
-def navigation(
-    config_dir: Annotated[
-        str, typer.Option("--config-dir", help="Configuration directory to use.")
-    ] = "config",
-    cycles: Annotated[
-        int, typer.Option("--cycles", help="How many full navigation cycles to run.")
-    ] = 2,
-    hold_seconds: Annotated[
-        float,
-        typer.Option(
-            "--hold-seconds",
-            help="How long to pump after each simulated click or route change.",
-        ),
-    ] = 0.35,
-    idle_seconds: Annotated[
-        float,
-        typer.Option(
-            "--idle-seconds",
-            help="How long to leave each exercised screen idle before the next action.",
-        ),
-    ] = 3.0,
-    tail_idle_seconds: Annotated[
-        float,
-        typer.Option(
-            "--tail-idle-seconds",
-            help="Final idle dwell on the hub after all navigation cycles complete.",
-        ),
-    ] = 10.0,
-    with_playback: Annotated[
-        bool,
-        typer.Option(
-            "--with-playback/--no-with-playback",
-            help="Drive playlist and shuffle playback paths during the soak.",
-        ),
-    ] = True,
-    provision_test_music: Annotated[
-        bool,
-        typer.Option(
-            "--provision-test-music/--no-provision-test-music",
-            help="Seed deterministic validation music before playback-driven navigation.",
-        ),
-    ] = True,
-    test_music_dir: Annotated[
-        str,
-        typer.Option(
-            "--test-music-dir",
-            help="Dedicated target directory for validation-only test music assets.",
-        ),
-    ] = DEFAULT_TEST_MUSIC_TARGET_DIR,
-    skip_sleep: Annotated[
-        bool, typer.Option("--skip-sleep", help="Skip the final sleep/wake exercise.")
-    ] = False,
-    verbose: Annotated[bool, typer.Option("--verbose", help="Enable DEBUG logging.")] = False,
-) -> None:
-    """Run the one-button target navigation and idle stability soak on LVGL hardware."""
-    from loguru import logger
-
-    configure_logging(verbose)
-
-    ok, details = run_navigation_soak(
-        config_dir=config_dir,
-        cycles=cycles,
-        hold_seconds=hold_seconds,
-        idle_seconds=idle_seconds,
-        tail_idle_seconds=tail_idle_seconds,
-        with_playback=with_playback,
-        provision_test_music=provision_test_music,
-        test_music_dir=test_music_dir,
-        skip_sleep=skip_sleep,
-    )
-    if ok:
-        logger.info("Navigation soak passed: {}", details)
-        return
-
-    logger.error("Navigation soak failed: {}", details)
-    raise typer.Exit(code=1)
+register_navigation_command(validate_app)
